Fixed scientific notation.

+++ NumberImput/doubleNumber.h
@@ -2,8 +2,8 @@
 #ifndef DOUBLENUMBER_H
 #define DOUBLENUMBER_H
 
-double userNumber();
-double doubleNumber(double x);
+int64_t userNumber();
+int64_t doubleNumber(int64_t x);
 void printStuff1();
 void printStuff();
